requirements.txt
xgboost==3.2.0
scikit-learn==1.7.2
numpy==2.2.6
pandas==2.3.3
optuna==4.7.0
optuna-integration==4.7.0
PyYAML==6.0.3
setuptools==65.5.0
